split off getGroups unit tests
